refactor: 타입 및 상수 분리
